chaos-tests/requirements.txt
# https://github.com/google/python-fire
fire
pyjavaproperties
pylint==2.4.4
PyYAML==5.3
pymongo==3.10.1